Do not increase active_tasks if the container is already present.
This fixes a bug where the active_tasks counter (used in the limit-active-tasks strategy)
would be increased even if the container running the task is already present in the worker.

The above bug can manifest, for instance, if the ATC restarts.

To be able to selectively increase the active_tasks counter when needed the FindOrChooseWorkerForContainer
in the Pool has been split in two so it can be known if the container was already present in the worker
and thus that worker has been chosen or if a new container has been spawned.

The context is removed from FindOrChooseWorkerForContainer as it is never used.