worker: implement backend.Ping
implementing `ping`, we're able to at least validate whether
`containerdRunnner()` gives back a runner that provides the proper
wiring.

for testing it in an unitary fashion, fakes are created for what could
be the `containerd` client (called `libcontainerd` here)

ps.: `Ping()` is currently used by `worker/heartbeater`, thus, it's an
endpoint that we definitely need to implement at some point anyway.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>