worker: make containerd gdn backend a restarter
given that there's a race between the moment that `containerd` gets up,
and the time at which our garden backend is able to communicate with it,
making the garden backend a `restart.Restarter` allows us to recover
until `containerd` is finally up.

ps.: in the near future, it might be wise to make this not restart
forever though (have some sort of time-based restart constraint).

pps.: here we also move the setting of PATH to `worker_linux.go` as
that's something expected by both gdn and containerd runners.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>