worker: add spec pkg for gdn->oci conversion
As we're now the ones implementing the low-level functionality of
directing containerd to create a container from a very high-level
description that `garden.ContainerSpec` gives, we need a way of coming
up with a valid OCI runtime spec to be used by containerd.

	be.create(garden.ContainerSpec{})
		|
		*--> containerd.newcontaiiner(id, oci_spec)

This commit introduces the `spec` package to deal with such conversion.

The details in this commit are mostly based on the current OCI spec
internal conversion that Guardian (`gdn`) performs (see references),
rather than taking the defaults that `containerd` comes with, as we're
trying to introduce the less amount of friction possible.

ref: https://github.com/opencontainers/runtime-spec/blob/master/spec.md
ref: https://github.com/cloudfoundry/guardian/blob/6b021168907b2/guardiancmd/server.go
ref: https://github.com/cloudfoundry/guardian/blob/0a658a3e51595/guardiancmd/command.go

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>