Label TasksWaiting metric in for Prometheus.
The `TasksWaiting` metric is now assigned labels for:

 * Team ID

 * Worker tags: when multiple tags are defined they are joined

 * Platform

of the tasks that are waiting to be executed.

Since one of the most interesting aspect of the metric is
to be able to AutoScale the workers instances, the added
labels allow for a more specific scaling since they
differentiate the groups the workers might belong to.

Signed-off-by: Alessandro Degano <alessandro.degano@pix4d.com>